+++ packages/client/src/Call.ts
@@ -1,5 +1,6 @@
 import { StreamSfuClient } from './StreamSfuClient';
 import {
+  BasePeerConnectionOpts,
   Dispatcher,
   getGenericSdp,
   isAudioTrackType,
@@ -25,6 +26,7 @@ import {
   createSubscription,
   getCurrentValue,
 } from './store/rxUtils';
+import { ScopedLogger, videoLoggerSystem } from './logger';
 import type {
   AcceptCallResponse,
   BlockUserRequest,
@@ -117,6 +119,7 @@ import {
   ClientDetails,
   Codec,
   ParticipantSource,
+  PeerType,
   PublishOption,
   SubscribeOption,
   TrackType,
@@ -139,12 +142,10 @@ import {
   AllCallEvents,
   CallEventListener,
   ErrorFromResponse,
-  Logger,
   RejectReason,
   StreamCallEvent,
 } from './coordinator/connection/types';
 import { getClientDetails } from './helpers/client-details';
-import { getLogger } from './logger';
 import {
   CameraManager,
   MicrophoneManager,
@@ -229,7 +230,7 @@ export class Call {
    */
   readonly permissionsContext = new PermissionsContext();
   readonly tracer = new Tracer(null);
-  readonly logger: Logger;
+  readonly logger: ScopedLogger;
 
   /**
    * The event dispatcher instance dedicated to this Call instance.
@@ -312,7 +313,7 @@ export class Call {
     this.streamClient = streamClient;
     this.clientStore = clientStore;
     this.streamClientBasePath = `/call/${this.type}/${this.id}`;
-    this.logger = getLogger(['Call']);
+    this.logger = videoLoggerSystem.getLogger('Call');
 
     const callTypeConfig = CallTypes.get(type);
     const participantSorter =
@@ -396,9 +397,9 @@ export class Call {
         if (!blockedUserIds || blockedUserIds.length === 0) return;
         const currentUserId = this.currentUserId;
         if (currentUserId && blockedUserIds.includes(currentUserId)) {
-          this.logger('info', 'Leaving call because of being blocked');
+          this.logger.info('Leaving call because of being blocked');
           await this.leave({ message: 'user blocked' }).catch((err) => {
-            this.logger('error', 'Error leaving call after being blocked', err);
+            this.logger.error('Error leaving call after being blocked', err);
           });
         }
       }),
@@ -442,8 +443,7 @@ export class Call {
           !hasPending(this.joinLeaveConcurrencyTag)
         ) {
           this.leave().catch(() => {
-            this.logger(
-              'error',
+            this.logger.error(
               'Could not leave a call that was accepted or rejected elsewhere',
             );
           });
@@ -526,8 +526,7 @@ export class Call {
             break;
         }
       } catch (err) {
-        this.logger(
-          'error',
+        this.logger.error(
           `Can't disable mic/camera/screenshare after revoked permissions`,
           err,
         );
@@ -894,12 +893,12 @@ export class Call {
     maxJoinRetries = Math.max(maxJoinRetries, 1);
     for (let attempt = 0; attempt < maxJoinRetries; attempt++) {
       try {
-        this.logger('trace', `Joining call (${attempt})`, this.cid);
+        this.logger.trace(`Joining call (${attempt})`, this.cid);
         await this.doJoin(data);
         delete joinData.migrating_from;
         break;
       } catch (err) {
-        this.logger('warn', `Failed to join call (${attempt})`, this.cid);
+        this.logger.warn(`Failed to join call (${attempt})`, this.cid);
         if (err instanceof ErrorFromResponse && err.unrecoverable) {
           // if the error is unrecoverable, we should not retry as that signals
           // that connectivity is good, but the coordinator doesn't allow the user
@@ -935,7 +934,7 @@ export class Call {
 
     this.joinCallData = data;
 
-    this.logger('debug', 'Starting join flow');
+    this.logger.debug('Starting join flow');
     this.state.setCallingState(CallingState.JOINING);
 
     const performingMigration =
@@ -998,9 +997,11 @@ export class Call {
       // prepare a generic SDP and send it to the SFU.
       // these are throw-away SDPs that the SFU will use to determine
       // the capabilities of the client (codec support, etc.)
+      const { dangerouslyForceCodec, fmtpLine, subscriberFmtpLine } =
+        this.clientPublishOptions || {};
       const [subscriberSdp, publisherSdp] = await Promise.all([
-        getGenericSdp('recvonly'),
-        getGenericSdp('sendonly'),
+        getGenericSdp('recvonly', dangerouslyForceCodec, subscriberFmtpLine),
+        getGenericSdp('sendonly', dangerouslyForceCodec, fmtpLine),
       ]);
       const isReconnecting =
         this.reconnectStrategy !== WebsocketReconnectStrategy.UNSPECIFIED;
@@ -1039,7 +1040,7 @@ export class Call {
           );
         }
       } catch (error) {
-        this.logger('warn', 'Join SFU request failed', error);
+        this.logger.warn('Join SFU request failed', error);
         sfuClient.close(
           StreamSfuClient.JOIN_FAILED,
           'Join request failed, connection considered unhealthy',
@@ -1111,7 +1112,7 @@ export class Call {
     this.reconnectStrategy = WebsocketReconnectStrategy.UNSPECIFIED;
     this.reconnectReason = '';
 
-    this.logger('info', `Joined call ${this.cid}`);
+    this.logger.info(`Joined call ${this.cid}`);
   };
 
   /**
@@ -1248,20 +1249,23 @@ export class Call {
     if (closePreviousInstances && this.subscriber) {
       this.subscriber.dispose();
     }
-    this.subscriber = new Subscriber({
+    const basePeerConnectionOptions: BasePeerConnectionOpts = {
       sfuClient,
       dispatcher: this.dispatcher,
       state: this.state,
       connectionConfig,
       tag: sfuClient.tag,
       enableTracing,
-      onReconnectionNeeded: (kind, reason) => {
+      clientPublishOptions: this.clientPublishOptions,
+      onReconnectionNeeded: (kind, reason, peerType) => {
         this.reconnect(kind, reason).catch((err) => {
-          const message = `[Reconnect] Error reconnecting after a subscriber error: ${reason}`;
-          this.logger('warn', message, err);
+          const message = `[Reconnect] Error reconnecting, after a ${PeerType[peerType]} error: ${reason}`;
+          this.logger.warn(message, err);
         });
       },
-    });
+    };
+
+    this.subscriber = new Subscriber(basePeerConnectionOptions);
 
     // anonymous users can't publish anything hence, there is no need
     // to create Publisher Peer Connection for them
@@ -1270,21 +1274,7 @@ export class Call {
       if (closePreviousInstances && this.publisher) {
         this.publisher.dispose();
       }
-      this.publisher = new Publisher({
-        sfuClient,
-        dispatcher: this.dispatcher,
-        state: this.state,
-        connectionConfig,
-        publishOptions,
-        tag: sfuClient.tag,
-        enableTracing,
-        onReconnectionNeeded: (kind, reason) => {
-          this.reconnect(kind, reason).catch((err) => {
-            const message = `[Reconnect] Error reconnecting after a publisher error: ${reason}`;
-            this.logger('warn', message, err);
-          });
-        },
-      });
+      this.publisher = new Publisher(basePeerConnectionOptions, publishOptions);
     }
 
     this.statsReporter?.stop();
@@ -1366,7 +1356,7 @@ export class Call {
     sfuClient: StreamSfuClient,
     reason: string,
   ) => {
-    this.logger('debug', '[Reconnect] SFU signal connection closed');
+    this.logger.debug('[Reconnect] SFU signal connection closed');
     const { callingState } = this.state;
     if (
       // SFU WS closed before we finished current join,
@@ -1389,7 +1379,7 @@ export class Call {
         ? WebsocketReconnectStrategy.FAST
         : WebsocketReconnectStrategy.REJOIN;
     this.reconnect(strategy, reason).catch((err) => {
-      this.logger('warn', '[Reconnect] Error reconnecting', err);
+      this.logger.warn('[Reconnect] Error reconnecting', err);
     });
   };
 
@@ -1435,8 +1425,7 @@ export class Call {
           reconnectingTime / 1000 > this.disconnectionTimeoutSeconds;
 
         if (shouldGiveUpReconnecting) {
-          this.logger(
-            'warn',
+          this.logger.warn(
             '[Reconnect] Stopping reconnection attempts after reaching disconnection timeout',
           );
           await markAsReconnectingFailed();
@@ -1453,16 +1442,14 @@ export class Call {
           // wait until the network is available
           await this.networkAvailableTask?.promise;
 
-          this.logger(
-            'info',
+          this.logger.info(
             `[Reconnect] Reconnecting with strategy ${WebsocketReconnectStrategy[this.reconnectStrategy]}`,
           );
 
           switch (this.reconnectStrategy) {
             case WebsocketReconnectStrategy.UNSPECIFIED:
             case WebsocketReconnectStrategy.DISCONNECT:
-              this.logger(
-                'debug',
+              this.logger.debug(
                 `[Reconnect] No-op strategy ${currentStrategy}`,
               );
               break;
@@ -1485,8 +1472,7 @@ export class Call {
           break; // do-while loop, reconnection worked, exit the loop
         } catch (error) {
           if (this.state.callingState === CallingState.OFFLINE) {
-            this.logger(
-              'debug',
+            this.logger.debug(
               `[Reconnect] Can't reconnect while offline, stopping reconnection attempts`,
             );
             break;
@@ -1494,8 +1480,7 @@ export class Call {
             // network change event will trigger the reconnection
           }
           if (error instanceof ErrorFromResponse && error.unrecoverable) {
-            this.logger(
-              'warn',
+            this.logger.warn(
               `[Reconnect] Can't reconnect due to coordinator unrecoverable error`,
               error,
             );
@@ -1528,8 +1513,7 @@ export class Call {
             : WebsocketReconnectStrategy.FAST;
           this.reconnectStrategy = nextStrategy;
 
-          this.logger(
-            'info',
+          this.logger.info(
             `[Reconnect] ${currentStrategy} (${this.reconnectAttempts}) failed. Attempting with ${WebsocketReconnectStrategy[nextStrategy]}`,
             error,
           );
@@ -1539,7 +1523,7 @@ export class Call {
         this.state.callingState !== CallingState.RECONNECTING_FAILED &&
         this.state.callingState !== CallingState.LEFT
       );
-      this.logger('info', '[Reconnect] Reconnection flow finished');
+      this.logger.info('[Reconnect] Reconnection flow finished');
     });
   };
 
@@ -1641,7 +1625,7 @@ export class Call {
     // handles the legacy "goAway" event
     const unregisterGoAway = this.on('goAway', () => {
       this.reconnect(WebsocketReconnectStrategy.MIGRATE, 'goAway').catch(
-        (err) => this.logger('warn', '[Reconnect] Error reconnecting', err),
+        (err) => this.logger.warn('[Reconnect] Error reconnecting', err),
       );
     });
 
@@ -1651,11 +1635,11 @@ export class Call {
       if (strategy === WebsocketReconnectStrategy.UNSPECIFIED) return;
       if (strategy === WebsocketReconnectStrategy.DISCONNECT) {
         this.leave({ message: 'SFU instructed to disconnect' }).catch((err) => {
-          this.logger('warn', `Can't leave call after disconnect request`, err);
+          this.logger.warn(`Can't leave call after disconnect request`, err);
         });
       } else {
         this.reconnect(strategy, error?.message || 'SFU Error').catch((err) => {
-          this.logger('warn', '[Reconnect] Error reconnecting', err);
+          this.logger.warn('[Reconnect] Error reconnecting', err);
         });
       }
     });
@@ -1665,7 +1649,7 @@ export class Call {
       (e) => {
         this.tracer.trace('network.changed', e);
         if (!e.online) {
-          this.logger('debug', '[Reconnect] Going offline');
+          this.logger.debug('[Reconnect] Going offline');
           if (!this.hasJoinedOnce) return;
           this.lastOfflineTimestamp = Date.now();
           // create a new task that would resolve when the network is available
@@ -1682,8 +1666,7 @@ export class Call {
             }
 
             this.reconnect(strategy, 'Going online').catch((err) => {
-              this.logger(
-                'warn',
+              this.logger.warn(
                 '[Reconnect] Error reconnecting after going online',
                 err,
               );
@@ -1693,7 +1676,7 @@ export class Call {
           this.sfuStatsReporter?.stop();
           this.state.setCallingState(CallingState.OFFLINE);
         } else {
-          this.logger('debug', '[Reconnect] Going online');
+          this.logger.debug('[Reconnect] Going online');
           this.sfuClient?.close(
             StreamSfuClient.DISPOSE_OLD_SOCKET,
             'Closing WS to reconnect after going online',
@@ -1883,14 +1866,12 @@ export class Call {
    * @param options the options to use.
    */
   updatePublishOptions = (options: ClientPublishOptions) => {
-    this.logger(
-      'warn',
+    this.logger.warn(
       '[call.updatePublishOptions]: You are manually overriding the publish options for this call. ' +
         'This is not recommended, and it can cause call stability/compatibility issues. Use with caution.',
     );
     if (this.state.callingState === CallingState.JOINED) {
-      this.logger(
-        'warn',
+      this.logger.warn(
         'Updating publish options after joining the call does not have an effect',
       );
     }
@@ -1904,7 +1885,7 @@ export class Call {
    */
   notifyNoiseCancellationStarting = async () => {
     return this.sfuClient?.startNoiseCancellation().catch((err) => {
-      this.logger('warn', 'Failed to notify start of noise cancellation', err);
+      this.logger.warn('Failed to notify start of noise cancellation', err);
     });
   };
 
@@ -1915,7 +1896,7 @@ export class Call {
    */
   notifyNoiseCancellationStopped = async () => {
     return this.sfuClient?.stopNoiseCancellation().catch((err) => {
-      this.logger('warn', 'Failed to notify stop of noise cancellation', err);
+      this.logger.warn('Failed to notify stop of noise cancellation', err);
     });
   };
 
@@ -2497,7 +2478,7 @@ export class Call {
         reason: 'timeout',
         message: `ringing timeout - ${this.isCreatedByMe ? 'no one accepted' : `user didn't interact with incoming call screen`}`,
       }).catch((err) => {
-        this.logger('error', 'Failed to drop call', err);
+        this.logger.error('Failed to drop call', err);
       });
     }, timeoutInMs);
   };
@@ -2661,10 +2642,10 @@ export class Call {
     publish: boolean,
   ) => {
     await this.camera.apply(settings.video, publish).catch((err) => {
-      this.logger('warn', 'Camera init failed', err);
+      this.logger.warn('Camera init failed', err);
     });
     await this.microphone.apply(settings.audio, publish).catch((err) => {
-      this.logger('warn', 'Mic init failed', err);
+      this.logger.warn('Mic init failed', err);
     });
   };