+++ packages/react-native-sdk/ios/StreamInCallManager.swift
@@ -96,6 +96,17 @@ class StreamInCallManager: RCTEventEmitter {
             if !audioManagerActivated {
                 return
             }
+            let session = RTCAudioSession.sharedInstance()
+            session.lockForConfiguration()
+            defer {
+                session.unlockForConfiguration()
+            }
+
+            do {
+                try session.setActive(false)
+            } catch {
+                log("Error deactivating audio session: \(error.localizedDescription)")
+            }
             if let prev = previousAudioSessionState {
                 let session = AVAudioSession.sharedInstance()
                 do {
@@ -129,13 +140,21 @@ class StreamInCallManager: RCTEventEmitter {
         } else {
             intendedCategory = .playAndRecord
             intendedMode = .voiceChat
+            
+            // XCode 16 and older don't expose .allowBluetoothHFP
+            // https://forums.swift.org/t/xcode-26-avaudiosession-categoryoptions-allowbluetooth-deprecated/80956
+            #if compiler(>=6.2) // For Xcode 26.0+
+                let bluetoothOption: AVAudioSession.CategoryOptions = .allowBluetoothHFP
+            #else
+                let bluetoothOption: AVAudioSession.CategoryOptions = .allowBluetooth
+            #endif
 
             if (defaultAudioDevice == .speaker) {
                 // defaultToSpeaker will route to speaker if nothing else is connected
-                intendedOptions = [.allowBluetoothHFP, .defaultToSpeaker]
+                intendedOptions = [bluetoothOption, .defaultToSpeaker]
             } else {
                 // having no defaultToSpeaker makes sure audio goes to earpiece if nothing is connected
-                intendedOptions = [.allowBluetoothHFP]
+                intendedOptions = [bluetoothOption]
             }
         }
 
@@ -147,32 +166,25 @@ class StreamInCallManager: RCTEventEmitter {
         RTCAudioSessionConfiguration.setWebRTC(rtcConfig)
         // END
 
-        // START: compare current audio session with intended, and update if different
+        // START: activate the webrtc audio session
         let session = RTCAudioSession.sharedInstance()
-        let currentCategory = session.category
-        let currentMode = session.mode
-        let currentOptions = session.categoryOptions
-        let currentIsActive = session.isActive
-
-        if currentCategory != intendedCategory.rawValue || currentMode != intendedMode.rawValue || currentOptions != intendedOptions || !currentIsActive {
-            session.lockForConfiguration()
-            do {
-                try session.setCategory(intendedCategory, mode: intendedMode, options: intendedOptions)
-                try session.setActive(true)
-                log("configureAudioSession: setCategory success \(intendedCategory.rawValue) \(intendedMode.rawValue) \(intendedOptions.rawValue)")
-            } catch {
-                log("configureAudioSession: setCategory failed due to: \(error.localizedDescription)")
-                do {
-                    try session.setMode(intendedMode)
-                    try session.setActive(true)
-                    log("configureAudioSession: setMode success \(intendedMode.rawValue)")
-                } catch {
-                    log("configureAudioSession: Error setting mode: \(error.localizedDescription)")
-                }
-            }
+        session.lockForConfiguration()
+        defer {
             session.unlockForConfiguration()
-        } else {
-            log("configureAudioSession: no change needed")
+        }
+        do {
+            try session.setCategory(intendedCategory, mode: intendedMode, options: intendedOptions)
+            try session.setActive(true)
+            log("configureAudioSession: setCategory success \(intendedCategory.rawValue) \(intendedMode.rawValue) \(intendedOptions.rawValue)")
+        } catch {
+            log("configureAudioSession: setCategory failed due to: \(error.localizedDescription)")
+            do {
+                try session.setMode(intendedMode)
+                try session.setActive(true)
+                log("configureAudioSession: setMode success \(intendedMode.rawValue)")
+            } catch {
+                log("configureAudioSession: Error setting mode: \(error.localizedDescription)")
+            }
         }
         // END
     }