+++ packages/react-native-sdk/src/contexts/internal/ScreenshotIosContext.tsx
@@ -1,7 +1,7 @@
 import {
   StreamVideoParticipant,
+  videoLoggerSystem,
   type VideoTrackType,
-  getLogger,
 } from '@stream-io/video-client';
 import React, {
   createContext,
@@ -85,20 +85,20 @@ export const ScreenshotIosContextProvider = ({
           `${participant.userId}-${videoTrackType}`,
         );
         if (!ref || !ref.current) {
-          getLogger(['ScreenshotIosContextProvider'])(
-            'error',
-            'Cannot take snapshot: No registered view for this participant',
-          );
+          videoLoggerSystem
+            .getLogger('ScreenshotIosContextProvider')
+            .error(
+              'Cannot take snapshot: No registered view for this participant',
+            );
           return null;
         }
 
         // Get the native handle for the view
         const tag = findNodeHandle(ref.current);
         if (!tag) {
-          getLogger(['ScreenshotIosContextProvider'])(
-            'error',
-            'Cannot take snapshot: Cannot get native handle for view',
-          );
+          videoLoggerSystem
+            .getLogger('ScreenshotIosContextProvider')
+            .error('Cannot take snapshot: Cannot get native handle for view');
           return null;
         }
 
@@ -107,11 +107,9 @@ export const ScreenshotIosContextProvider = ({
 
         return base64Image;
       } catch (error) {
-        getLogger(['ScreenshotIosContextProvider'])(
-          'error',
-          'Error taking participant snapshot:',
-          error,
-        );
+        videoLoggerSystem
+          .getLogger('ScreenshotIosContextProvider')
+          .error('Error taking participant snapshot:', error);
         return null;
       }
     },