+++ packages/react-native-sdk/src/providers/NoiseCancellation/NoiseCancellationProvider.tsx
@@ -6,9 +6,9 @@ import React, {
   useState,
 } from 'react';
 import {
-  getLogger,
   NoiseCancellationSettingsModeEnum,
   OwnCapability,
+  videoLoggerSystem,
 } from '@stream-io/video-client';
 
 import { useCall, useCallStateHooks } from '@stream-io/video-react-bindings';
@@ -100,22 +100,18 @@ export const NoiseCancellationProvider = (props: PropsWithChildren<{}>) => {
     call.microphone
       .enableNoiseCancellation(ncInstance)
       .catch((err) =>
-        getLogger(['NoiseCancellationProvider'])(
-          'error',
-          `Can't initialize noise suppression`,
-          err,
-        ),
+        videoLoggerSystem
+          .getLogger('NoiseCancellationProvider')
+          .error(`Can't initialize noise suppression`, err),
       );
 
     return () => {
       call.microphone
         .disableNoiseCancellation()
         .catch((err) =>
-          getLogger(['NoiseCancellationProvider'])(
-            'error',
-            `Can't disable noise suppression`,
-            err,
-          ),
+          videoLoggerSystem
+            .getLogger('NoiseCancellationProvider')
+            .error(`Can't disable noise suppression`, err),
         );
       unsubscribe();
     };